+++ example/android/app/src/main/kotlin/flutter/soloud/flutter_soloud_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package flutter.soloud.flutter_soloud_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()