+++ example/lib/buffer_stream/websocket.dart
@@ -42,7 +42,7 @@ void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   /// Initialize the player.
-  await SoLoud.instance.init();
+  await SoLoud.instance.init(sampleRate: 24000);
 
   runApp(
     const MaterialApp(
@@ -61,11 +61,17 @@ class WebsocketExample extends StatefulWidget {
 
 class _WebsocketExampleState extends State<WebsocketExample> {
   final websocketUri = 'ws://192.168.1.2:8080/';
-  final sampleRate = [11025, 22050, 44100, 48000];
-  final format = ['f32le', 's8', 's16le', 's32le'];
-  int srId = 2;
+  // Supported Opus sample rates:
+  // 48000
+  // 24000
+  // 16000
+  // 12000
+  // 8000
+  final sampleRate = [8000, 12000, 16000, 24000, 44100, 48000];
+  final format = ['f32le', 's8', 's16le', 's32le', 'opus'];
+  int srId = 5;
   int chId = 0;
-  int fmtId = 0;
+  int fmtId = 4;
   WebSocketChannel? channel;
   AudioSource? currentSound;
   SoundHandle? handle;
@@ -183,10 +189,11 @@ class _WebsocketExampleState extends State<WebsocketExample> {
             onPressed: () async {
               await channel?.sink.close();
               await SoLoud.instance.disposeAllSources();
+              streamBuffering.value = false;
 
               currentSound = SoLoud.instance.setBufferStream(
-                maxBufferSize: 1024 * 1024 * 600, // 150 MB
-                bufferingTimeNeeds: 3,
+                maxBufferSize: 1024 * 1024 * 200, // 200 MB
+                bufferingTimeNeeds: 0.5,
                 sampleRate: sampleRate[srId],
                 channels: Channels.values[chId],
                 format: BufferType.values[fmtId],
@@ -244,6 +251,11 @@ class _WebsocketExampleState extends State<WebsocketExample> {
                     debugPrint('error adding audio data: $e');
                     await channel?.sink.close();
                   }
+
+                  // start playing at first audio chunk received
+                  if (numberOfChunks == 1) {
+                    handle = await SoLoud.instance.play(currentSound!);
+                  }
                 },
                 onDone: () {
                   if (currentSound != null) {
@@ -271,7 +283,6 @@ class _WebsocketExampleState extends State<WebsocketExample> {
                     volume: 0.6,
                     // looping: true,
                   );
-                  print('handle: $handle');
                   Timer.periodic(const Duration(milliseconds: 1000), (timer) {
                     if (currentSound == null ||
                         SoLoud.instance.getIsValidVoiceHandle(handle!) ==