+++ lib/src/soloud.dart
@@ -168,6 +168,9 @@ interface class SoLoud {
   /// A helper for loading files that aren't on disk.
   final SoLoudLoader _loader = SoLoudLoader();
 
+  /// Wheter or not the Opus and Ogg libraries are available.
+  bool _areOpusOggLibsAvailable = false;
+
   /// The backing private field for [isInitialized].
   // TODO(filip): related to `get initialized`
   // ignore: prefer_final_fields
@@ -338,10 +341,12 @@ interface class SoLoud {
       deinit();
     }
 
+    _areOpusOggLibsAvailable = _controller.soLoudFFI.areOpusOggLibsAvailable();
+
     _activeSounds.clear();
 
     // Initialize native callbacks
-    _initializeNativeCallbacks();
+    await _initializeNativeCallbacks();
 
     final error = _controller.soLoudFFI.initEngine(
       device?.id ?? -1,
@@ -432,9 +437,9 @@ interface class SoLoud {
   // TODO(filip): 'setDartEventCallbacks()' can be called more then once,
   // please take a look at the listeners if you find a better way
   // to manage them only once.
-  void _initializeNativeCallbacks() {
+  Future<void> _initializeNativeCallbacks() async {
     // Initialize callbacks.
-    _controller.soLoudFFI.setDartEventCallbacks();
+    await _controller.soLoudFFI.setDartEventCallbacks();
 
     // Listen when a handle becomes invalid because has been stopped/ended.
     if (!_controller.soLoudFFI.voiceEndedEventController.hasListener) {
@@ -672,10 +677,18 @@ interface class SoLoud {
   /// waiting until the buffer will have enough data to cover this time.
   ///
   /// [sampleRate] the sample rate. Usually is 22050 or 44100 (CD quality).
+  /// When using [format] as `opus`, the sample rate can be 48000, 24000,
+  /// 16000, 12000 or 8000. Whatever the sample rate of the incoming data is,
+  /// it will be resampled to this value. So, if you are adding Opus data at
+  /// 48 KHz, and you set this to 24000, the data will be resampled to 24 KHz.
   ///
-  /// [channels] enum to choose the number of channels.
+  /// [channels] enum to choose the number of channels. The `opus` format
+  /// supports only mono and stereo.
   ///
-  /// [format] enum to choose from `f32le`, `s8`, `s16le` and `s32le`.
+  /// [format] enum to choose from `f32le`, `s8`, `s16le`, `s32le` and
+  /// `opus`. The last one is a special format that uses the Opus codec with
+  /// Ogg container. It supports only 48, 24, 16, 12 and 8 KHz sample rates
+  /// and mono and stereo.
   ///
   /// [onBuffering] a callback that is called when starting to buffer
   /// (isBuffering = true) and when the buffering is done (isBuffering = false).
@@ -683,16 +696,52 @@ interface class SoLoud {
   /// the `time` in seconds.
   ///
   /// Throws [SoLoudNotInitializedException] if the engine is not initialized.
+  /// Throws [SoLoudOpusOggLibsNotAvailableException] if trying to use the
+  /// `opus` format but the Opus and Ogg libraries are not available. Please
+  /// check the `README.md` file for more information.
   AudioSource setBufferStream({
     int maxBufferSize = 1024 * 1024 * 100, // 100 MB in bytes
     double bufferingTimeNeeds = 2, // 2 seconds of data needed to un-pause
-    int sampleRate = 22050,
+    int sampleRate = 24000,
     Channels channels = Channels.mono,
     BufferType format = BufferType.s16le,
     void Function(bool isBuffering, int handle, double time)? onBuffering,
   }) {
     if (!isInitialized) {
       throw const SoLoudNotInitializedException();
+    }
+
+    if (!_areOpusOggLibsAvailable && format == BufferType.opus) {
+      throw const SoLoudOpusOggLibsNotAvailableException();
+    }
+
+    final opusA = () {
+      if (format == BufferType.opus) {
+        return sampleRate == 48000 ||
+            sampleRate == 24000 ||
+            sampleRate == 16000 ||
+            sampleRate == 12000 ||
+            sampleRate == 8000;
+      }
+      return true;
+    }();
+    final opusB = () {
+      if (format == BufferType.opus) {
+        return channels == Channels.mono || channels == Channels.stereo;
+      }
+      return true;
+    }();
+    assert(
+      opusA,
+      'Opus format only supports 48, 24, 16, 12 and 8 KHz sample rates',
+    );
+    assert(
+      opusB,
+      'Only mono and stereo channels are supported for Opus format',
+    );
+
+    if (!opusA || !opusB) {
+      throw const SoLoudWrongOpusParamsException();
     }
 
     final ret = SoLoudController().soLoudFFI.setBufferStream(
@@ -703,6 +752,11 @@ interface class SoLoud {
           format.value,
           onBuffering,
         );
+
+    if (ret.error != PlayerErrors.noError) {
+      _logPlayerError(ret.error, from: 'addAudioDataStream() result');
+      throw SoLoudCppException.fromPlayerError(ret.error);
+    }
 
     final newSound = _addNewSound(ret.error, '', ret.soundHash.hash);
     return newSound;
@@ -731,7 +785,7 @@ interface class SoLoud {
   ///   final pcmBuffer = Uint8List(1024 * 1024); // 1 MB in bytes
   ///   final pcmAudio = SoLoud.instance.setBufferStream(
   ///     maxBufferSize: 1024 * 1024, // 1 MB in bytes
-  ///     pcmFormat: BufferPcmType.s8, // signed 8 bits
+  ///     format: BufferPcmType.s8, // signed 8 bits
   ///   );
   ///   for (var i = 0; i < pcmBuffer.length; i++) {
   ///     // Compose your PCM data here.